# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/members/participants": {
      "get": {
        "operationId": "MembersController_findParticipants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ParticipantDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      }
    },
    "/members": {
      "get": {
        "operationId": "MembersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      },
      "post": {
        "operationId": "MembersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemberDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      }
    },
    "/members/{id}": {
      "get": {
        "operationId": "MembersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      },
      "patch": {
        "operationId": "MembersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      },
      "delete": {
        "operationId": "MembersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "members"
        ]
      }
    },
...
    "/validation/submissions/{id}/evidences/{evidenceId}/url": {
      "get": {
        "operationId": "PlatformValidationsController_evidenceUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "evidenceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "platform-validations"
        ]
      }
    },
    "/submissions/{submissionId}/evidences": {
      "post": {
        "operationId": "EvidencesController_upload",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Evidence"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "evidences"
        ]
      }
    },
    "/submissions/{submissionId}/evidences/{evidenceId}/url": {
      "get": {
        "operationId": "EvidencesController_signedUrl",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "evidenceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "evidences"
        ]
      }
    },
    "/submissions/{submissionId}/evidences/{evidenceId}": {
      "delete": {
        "operationId": "EvidencesController_remove",
        "parameters": [
          {
            "name": "submissionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "evidenceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "evidences"
        ]
      }
    },
    "/goals": {
      "get": {
        "operationId": "GoalsController_findAll",
        "parameters": [
          {
            "name": "campaignId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Goal"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      },
      "post": {
        "operationId": "GoalsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      }
    },
    "/goals/monthly-plan": {
      "post": {
        "operationId": "GoalsController_createMonthlyPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMonthlyPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Goal"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      }
    },
    "/goals/{id}/progress": {
      "get": {
        "operationId": "GoalsController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalProgressDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      }
    },
    "/goals/{id}": {
      "get": {
        "operationId": "GoalsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      },
      "patch": {
        "operationId": "GoalsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGoalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Goal"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      },
      "delete": {
        "operationId": "GoalsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "goals"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gincana Solidária API",
    "description": "Multi-tenant API for solidarity campaigns, submissions and scoring.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Membership identifier"
          },
          "name": {
            "type": "string"
          },
          "status": {
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ],
            "type": "string",
            "example": "ACTIVE"
          }
        },
        "required": [
          "id",
          "name",
          "status"
        ]
      },
      "Membership": {
        "type": "object",
        "properties": {}
      },
      "CreateMemberDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "enum": [
              "MANAGER",
              "MEMBER"
            ],
            "type": "string"
          },
          "temporaryPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          }
        },
        "required": [
          "name",
          "email",
          "role",
          "temporaryPassword"
        ]
      },
      "UpdateMemberDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "enum": [
              "MANAGER",
              "MEMBER"
            ],
            "type": "string"
          },
          "temporaryPassword": {
            "type": "string",
            "minLength": 6,
            "maxLength": 128
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          }
        }
      },
...
      "Evidence": {
        "type": "object",
        "properties": {}
      },
      "Goal": {
        "type": "object",
        "properties": {}
      },
      "CreateMonthlyPlanDto": {
        "type": "object",
        "properties": {
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "format": "uuid"
          },
          "titlePrefix": {
            "type": "string",
            "maxLength": 140,
            "example": "Plano mensal"
          },
          "targetPoints": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetActions": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetParticipants": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetQuantity": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "campaignId",
          "titlePrefix"
        ]
      },
      "GoalMetricDto": {
        "type": "object",
        "properties": {
          "points": {
            "type": "number"
          },
          "actions": {
            "type": "number"
          },
          "participants": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "points",
          "actions",
          "participants",
          "quantity"
        ]
      },
      "GoalProgressDto": {
        "type": "object",
        "properties": {
          "achieved": {
            "$ref": "#/components/schemas/GoalMetricDto"
          },
          "targets": {
            "$ref": "#/components/schemas/GoalMetricDto"
          },
          "remaining": {
            "$ref": "#/components/schemas/GoalMetricDto"
          },
          "percentages": {
            "$ref": "#/components/schemas/GoalMetricDto"
          },
          "overallPercentage": {
            "type": "number"
          },
          "status": {
            "enum": [
              "NOT_STARTED",
              "IN_PROGRESS",
              "ACHIEVED",
              "EXPIRED"
            ],
            "type": "string"
          }
        },
        "required": [
          "achieved",
          "targets",
          "remaining",
          "percentages",
          "overallPercentage",
          "status"
        ]
      },
      "CreateGoalDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "type": {
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "CAMPAIGN",
              "CUSTOM"
            ],
            "type": "string"
          },
          "startsAt": {
            "type": "string",
            "format": "date"
          },
          "endsAt": {
            "type": "string",
            "format": "date"
          },
          "targetPoints": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetActions": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetParticipants": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetQuantity": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "title",
          "campaignId",
          "type",
          "startsAt",
          "endsAt"
        ]
      },
      "UpdateGoalDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 180
          },
          "description": {
            "type": "string"
          },
          "campaignId": {
            "type": "string",
            "format": "uuid"
          },
          "activityId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "type": {
            "enum": [
              "WEEKLY",
              "MONTHLY",
              "CAMPAIGN",
              "CUSTOM"
            ],
            "type": "string"
          },
          "startsAt": {
            "type": "string",
            "format": "date"
          },
          "endsAt": {
            "type": "string",
            "format": "date"
          },
          "targetPoints": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetActions": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetParticipants": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "targetQuantity": {
            "type": "number",
            "minimum": 0,
            "default": 0
          },
          "unit": {
            "type": "string",
            "maxLength": 40
          }
        }
      },
...
    }
  }
}